Split the FirmwareUpdater into blocking and async
